+++ aws-cdn-templates/coll/env-cdn.sh
@@ -1,6 +1,0 @@
-export ZONE_ID="Z0804276GZ3UW6RKRQ2B"
-export PORTALE_PA_CERTIFICATE_ARN="arn:aws:acm:us-east-1:615714398925:certificate/22f183e5-2745-467b-91b3-1594ea96f006"
-export PORTALE_PF_CERTIFICATE_ARN="arn:aws:acm:us-east-1:615714398925:certificate/0e4e8803-88b3-4071-9aed-58568214e0ed"
-export PORTALE_PF_LOGIN_CERTIFICATE_ARN="arn:aws:acm:us-east-1:615714398925:certificate/ece097ba-e079-4d61-b516-76ac6000e7a2"
-export PORTALE_PG_CERTIFICATE_ARN="arn:aws:acm:us-east-1:615714398925:certificate/180dbfb9-72e3-4744-afc4-5fee8ae62669"
-export REACT_APP_URL_API="https://webapi.coll.pn.pagopa.it/ https://s3.eu-south-1.amazonaws.com/pnsafestorage-nonlegal-input-eu-south-1/"