+++ packages/pn-personafisica-webapp/src/components/Contacts/SmsContactItem.tsx
@@ -38,6 +38,7 @@ import DeleteDialog from './DeleteDialog';
 import DigitalContact from './DigitalContact';
 import ExistingContactDialog from './ExistingContactDialog';
 import InformativeDialog from './InformativeDialog';
+import SpecialContacts from './SpecialContacts';
 
 enum ModalType {
   EXISTING = 'existing',
@@ -228,19 +229,21 @@ const SmsContactElem: React.FC<SmsElemProps> = ({ onCancelInsert, slotsProps, sl
 const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const dispatch = useAppDispatch();
-  const { defaultSERCQ_SENDAddress, defaultSMSAddress, addresses } = useAppSelector(
-    contactsSelectors.selectAddresses
-  );
+  const { defaultSERCQ_SENDAddress, defaultSMSAddress, addresses, specialSMSAddresses } =
+    useAppSelector(contactsSelectors.selectAddresses);
 
   const [modalOpen, setModalOpen] = useState<ModalType | null>(null);
   const [insertMode, setInsertMode] = useState(false);
 
   const isActive = !!defaultSMSAddress;
+  const blockDelete = specialSMSAddresses.length > 0;
 
   const hasCourtesyAddresses =
     addresses.filter(
       (addr) => addr.addressType === AddressType.COURTESY && addr.value !== IOAllowedValues.DISABLED
     ).length > 0;
+
+  const showSpecialContactsSection = specialSMSAddresses.length > 0;
 
   const deleteConfirmHandler = () => {
     setModalOpen(null);
@@ -275,6 +278,9 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
   };
 
   const getRemoveModalTitle = () => {
+    if (blockDelete) {
+      return t('courtesy-contacts.block-remove-sms-title', { ns: 'recapiti' });
+    }
     if (defaultSERCQ_SENDAddress) {
       return t(`courtesy-contacts.remove-sms-title-dod-enabled`, {
         ns: 'recapiti',
@@ -284,6 +290,9 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
   };
 
   const getRemoveModalMessage = () => {
+    if (blockDelete) {
+      return t('courtesy-contacts.block-remove-sms-message', { ns: 'recapiti' });
+    }
     if (defaultSERCQ_SENDAddress) {
       return (
         <Trans
@@ -357,6 +366,9 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
         >
           {t('courtesy-contacts.sms-description', { ns: 'recapiti' })}
         </Typography>
+        {showSpecialContactsSection && (
+          <SpecialContacts addressType={AddressType.COURTESY} channelType={ChannelType.SMS} />
+        )}
         <DeleteDialog
           showModal={modalOpen === ModalType.DELETE}
           removeModalTitle={getRemoveModalTitle()}
@@ -375,6 +387,7 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
                 : undefined,
             },
           }}
+          blockDelete={blockDelete}
         />
       </PnInfoCard>
     );