+++ packages/pn-personafisica-webapp/src/components/Contacts/__test__/CourtesyContacts.test.tsx
@@ -1,7 +1,23 @@
 import { digitalCourtesyAddresses } from '../../../__mocks__/Contacts.mock';
-import { fireEvent, getByText, render, waitFor, within } from '../../../__test__/test-utils';
+import {
+  fireEvent,
+  getByText,
+  render,
+  screen,
+  testStore,
+  waitFor,
+  within,
+} from '../../../__test__/test-utils';
 import { ChannelType } from '../../../models/contacts';
 import CourtesyContacts from '../CourtesyContacts';
+
+const defaultEmailAddress = digitalCourtesyAddresses.find(
+  (addr) => addr.senderId === 'default' && addr.channelType === ChannelType.EMAIL
+);
+
+const defaultSmsAddress = digitalCourtesyAddresses.find(
+  (addr) => addr.senderId === 'default' && addr.channelType === ChannelType.SMS
+);
 
 describe('CourtesyContacts Component', async () => {
   it('renders component - no contacts', async () => {
@@ -38,16 +54,10 @@ describe('CourtesyContacts Component', async () => {
     const { getByText, getByTestId } = render(<CourtesyContacts />, {
       preloadedState: { contactsState: { digitalAddresses: digitalCourtesyAddresses } },
     });
-    const defaultPhone = digitalCourtesyAddresses.find(
-      (addr) => addr.channelType === ChannelType.SMS && addr.senderId === 'default'
-    );
-    const defaultEmail = digitalCourtesyAddresses.find(
-      (addr) => addr.channelType === ChannelType.EMAIL && addr.senderId === 'default'
-    );
     const emailContact = getByTestId(`default_emailContact`);
     const emailInput = emailContact.querySelector(`[name="default_email"]`);
     expect(emailInput).not.toBeInTheDocument();
-    const email = getByText(defaultEmail!.value);
+    const email = getByText(defaultEmailAddress!.value);
     expect(email).toBeInTheDocument();
     const emailButton = within(emailContact).getByRole('button');
     expect(emailButton).toBeEnabled();
@@ -59,7 +69,7 @@ describe('CourtesyContacts Component', async () => {
     const phoneContact = getByTestId(`default_smsContact`);
     const phoneInput = phoneContact.querySelector(`[name="default_sms"]`);
     expect(phoneInput).not.toBeInTheDocument();
-    const phoneNumber = getByText(defaultPhone!.value);
+    const phoneNumber = getByText(defaultSmsAddress!.value);
     expect(phoneNumber).toBeInTheDocument();
     const phoneButton = within(phoneContact).getByRole('button');
     expect(phoneButton).toBeEnabled();
@@ -67,5 +77,86 @@ describe('CourtesyContacts Component', async () => {
     const smsDisableBtn = getByTestId('disable-sms');
     expect(smsDisableBtn).toBeInTheDocument();
     expect(smsDisableBtn).toHaveTextContent('disable');
+  });
+
+  it('renders component - Email enabled', async () => {
+    // render component
+    const { getByText, getByTestId } = render(<CourtesyContacts />, {
+      preloadedState: { contactsState: { digitalAddresses: digitalCourtesyAddresses } },
+    });
+    const emailContact = getByTestId(`default_emailContact`);
+    expect(emailContact).toBeInTheDocument();
+    const emailInput = emailContact.querySelector(`[name="default_email"]`);
+    expect(emailInput).not.toBeInTheDocument();
+    const email = getByText(defaultEmailAddress!.value);
+    expect(email).toBeInTheDocument();
+    const emailButton = within(emailContact).getByRole('button');
+    expect(emailButton).toBeEnabled();
+    expect(emailButton.textContent).toMatch('button.modifica');
+
+    const disableBtn = getByTestId('disable-email');
+    expect(disableBtn).toBeInTheDocument();
+    expect(disableBtn).toHaveTextContent('disable');
+
+    // verify email could not be disabled
+    fireEvent.click(disableBtn);
+    const dialog = await waitFor(() => screen.getByRole('dialog'));
+    expect(dialog).toBeInTheDocument();
+
+    const dialogTitle = getByText('courtesy-contacts.block-remove-email-title');
+    expect(dialogTitle).toBeInTheDocument();
+    const dialogMessage = getByText('courtesy-contacts.block-remove-email-message');
+    expect(dialogMessage).toBeInTheDocument();
+
+    const confirmBtn = screen.getByRole('button', { name: 'button.understand' });
+    fireEvent.click(confirmBtn);
+    await waitFor(() => {
+      expect(dialog).not.toBeInTheDocument();
+    });
+
+    expect(testStore.getState().contactsState.digitalAddresses).toStrictEqual(
+      digitalCourtesyAddresses
+    );
+  });
+
+  it('renders component - Sms enabled', async () => {
+    // render component
+    const { getByText, getByTestId } = render(<CourtesyContacts />, {
+      preloadedState: { contactsState: { digitalAddresses: digitalCourtesyAddresses } },
+    });
+    const smsContact = getByTestId(`default_smsContact`);
+    expect(smsContact).toBeInTheDocument();
+    const smsInput = smsContact.querySelector(`[name="default_sms"]`);
+    expect(smsInput).not.toBeInTheDocument();
+    const sms = getByText(defaultSmsAddress!.value);
+    expect(sms).toBeInTheDocument();
+    const smsButton = within(smsContact).getByRole('button');
+    expect(smsButton).toBeEnabled();
+    expect(smsButton.textContent).toMatch('button.modifica');
+
+    const disableBtn = getByTestId('disable-sms');
+    expect(disableBtn).toBeInTheDocument();
+    expect(disableBtn).toHaveTextContent('disable');
+    expect(disableBtn).toBeInTheDocument();
+
+    // verify email could not be disabled
+    fireEvent.click(disableBtn);
+    const dialog = await waitFor(() => screen.getByRole('dialog'));
+    expect(dialog).toBeInTheDocument();
+
+    const dialogTitle = getByText('courtesy-contacts.block-remove-sms-title');
+    expect(dialogTitle).toBeInTheDocument();
+    const dialogMessage = getByText('courtesy-contacts.block-remove-sms-message');
+    expect(dialogMessage).toBeInTheDocument();
+
+    const confirmBtn = screen.getByRole('button', { name: 'button.understand' });
+    fireEvent.click(confirmBtn);
+    await waitFor(() => {
+      expect(dialog).not.toBeInTheDocument();
+    });
+
+    expect(testStore.getState().contactsState.digitalAddresses).toStrictEqual(
+      digitalCourtesyAddresses
+    );
   });
 });