+++ packages/pn-personafisica-webapp/src/components/Contacts/__test__/DigitalContactActivation.test.tsx
@@ -270,6 +270,13 @@ describe('DigitalContactActivation', () => {
     const button = result.getByTestId('default_email-button');
     fireEvent.click(button);
 
+    // informative dialog
+    await waitFor(() => {
+      expect(result.getByRole('dialog')).toBeInTheDocument();
+    });
+    const infoButton = result.getByRole('button', { name: 'button.understand' });
+    fireEvent.click(infoButton);
+
     // inser otp and confirm
     const dialog = await fillCodeDialog(result);