+++ packages/pn-personagiuridica-webapp/src/components/Contacts/SmsContactItem.tsx
@@ -24,6 +24,7 @@ import DeleteDialog from './DeleteDialog';
 import DigitalContact from './DigitalContact';
 import ExistingContactDialog from './ExistingContactDialog';
 import InformativeDialog from './InformativeDialog';
+import SpecialContacts from './SpecialContacts';
 
 enum ModalType {
   EXISTING = 'existing',
@@ -191,17 +192,19 @@ const SmsContactElem: React.FC<SmsElemProps> = ({ onCancelInsert, slotsProps, sl
 const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const dispatch = useAppDispatch();
-  const { defaultSERCQ_SENDAddress, defaultSMSAddress, addresses } = useAppSelector(
-    contactsSelectors.selectAddresses
-  );
+  const { defaultSERCQ_SENDAddress, defaultSMSAddress, addresses, specialSMSAddresses } =
+    useAppSelector(contactsSelectors.selectAddresses);
 
   const [modalOpen, setModalOpen] = useState<ModalType | null>(null);
   const [insertMode, setInsertMode] = useState(false);
 
   const isActive = !!defaultSMSAddress;
+  const blockDelete = specialSMSAddresses.length > 0;
 
   const hasCourtesyAddresses =
     addresses.filter((addr) => addr.addressType === AddressType.COURTESY).length > 0;
+
+  const showSpecialContactsSection = specialSMSAddresses.length > 0;
 
   const deleteConfirmHandler = () => {
     setModalOpen(null);
@@ -235,6 +238,9 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
   };
 
   const getRemoveModalTitle = () => {
+    if (blockDelete) {
+      return t('courtesy-contacts.block-remove-sms-title', { ns: 'recapiti' });
+    }
     if (defaultSERCQ_SENDAddress) {
       return t(`courtesy-contacts.remove-sms-title-dod-enabled`, {
         ns: 'recapiti',
@@ -244,6 +250,9 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
   };
 
   const getRemoveModalMessage = () => {
+    if (blockDelete) {
+      return t('courtesy-contacts.block-remove-sms-message', { ns: 'recapiti' });
+    }
     if (defaultSERCQ_SENDAddress) {
       return (
         <Trans
@@ -317,6 +326,9 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
         >
           {t('courtesy-contacts.sms-description', { ns: 'recapiti' })}
         </Typography>
+        {showSpecialContactsSection && (
+          <SpecialContacts addressType={AddressType.COURTESY} channelType={ChannelType.SMS} />
+        )}
         <DeleteDialog
           showModal={modalOpen === ModalType.DELETE}
           removeModalTitle={getRemoveModalTitle()}
@@ -335,6 +347,7 @@ const SmsContactItem: React.FC<SmsItemProps> = ({ slotsProps, slots }) => {
                 : undefined,
             },
           }}
+          blockDelete={blockDelete}
         />
       </PnInfoCard>
     );